+++ wasm/parser/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) until v1.0.0.
 
+## [0.53.0] - 2025-02-26
+
+### Features
+
+- 835ee95 wasm: Return estree with utf16 span offsets (#9376) (Boshen)
+
+### Refactor
+
+- b09249c ast/estree: Rename serializers and serialization methods (#9284) (overlookmotel)
+
 ## [0.52.0] - 2025-02-21
 
 - 216b33f ast/estree: [**BREAKING**] Replace `serde` with custom `ESTree` serializer (#9256) (overlookmotel)